refactor: move ItemGrid.css to styles folder
